Refactor: move MainForm to its own file and add InitializeComponent for Windows Forms compatibility
